+++ src/coreclr/vm/crossgen/CMakeLists.txt
@@ -7,6 +7,7 @@ set(VM_CROSSGEN_SOURCES
   ../baseassemblyspec.cpp
   ../binder.cpp
   ../bundle.cpp
+  ../callconvbuilder.cpp
   ../castcache.cpp
   ../ceeload.cpp
   ../ceemain.cpp
