+++ src/coreclr/vm/dllimportcallback.cpp
@@ -14,15 +14,11 @@
 #include "object.h"
 #include "dllimportcallback.h"
 #include "mlinfo.h"
-#include "comdelegate.h"
 #include "ceeload.h"
 #include "eeconfig.h"
 #include "dbginterface.h"
 #include "stubgen.h"
 #include "appdomain.inl"
-#include "callingconvention.h"
-#include "customattribute.h"
-#include "typeparse.h"
 
 #ifndef CROSSGEN_COMPILE
 
@@ -554,99 +550,6 @@ void STDCALL LogUMTransition(UMEntryThunk* thunk)
     END_ENTRYPOINT_VOIDRET;
 
     }
-#endif
+#endif // _DEBUG
 
-bool TryGetCallingConventionFromUnmanagedCallersOnly(MethodDesc* pMD, CorInfoCallConvExtension* pCallConv)
-{
-    STANDARD_VM_CONTRACT;
-    _ASSERTE(pMD != NULL && pMD->HasUnmanagedCallersOnlyAttribute());
-
-    // Validate usage
-    COMDelegate::ThrowIfInvalidUnmanagedCallersOnlyUsage(pMD);
-
-    BYTE* pData = NULL;
-    LONG cData = 0;
-
-    bool nativeCallableInternalData = false;
-    HRESULT hr = pMD->GetCustomAttribute(WellKnownAttribute::UnmanagedCallersOnly, (const VOID **)(&pData), (ULONG *)&cData);
-    if (hr == S_FALSE)
-    {
-        hr = pMD->GetCustomAttribute(WellKnownAttribute::NativeCallableInternal, (const VOID **)(&pData), (ULONG *)&cData);
-        nativeCallableInternalData = SUCCEEDED(hr);
-    }
-
-    IfFailThrow(hr);
-
-    _ASSERTE(cData > 0);
-
-    CustomAttributeParser ca(pData, cData);
-
-    // UnmanagedCallersOnly and NativeCallableInternal each
-    // have optional named arguments.
-    CaNamedArg namedArgs[2];
-
-    // For the UnmanagedCallersOnly scenario.
-    CaType caCallConvs;
-
-    // Define attribute specific optional named properties
-    if (nativeCallableInternalData)
-    {
-        namedArgs[0].InitI4FieldEnum("CallingConvention", "System.Runtime.InteropServices.CallingConvention", (ULONG)(CorPinvokeMap)0);
-    }
-    else
-    {
-        caCallConvs.Init(SERIALIZATION_TYPE_SZARRAY, SERIALIZATION_TYPE_TYPE, SERIALIZATION_TYPE_UNDEFINED, NULL, 0);
-        namedArgs[0].Init("CallConvs", SERIALIZATION_TYPE_SZARRAY, caCallConvs);
-    }
-
-    // Define common optional named properties
-    CaTypeCtor caEntryPoint(SERIALIZATION_TYPE_STRING);
-    namedArgs[1].Init("EntryPoint", SERIALIZATION_TYPE_STRING, caEntryPoint);
-
-    InlineFactory<SArray<CaValue>, 4> caValueArrayFactory;
-    DomainAssembly* domainAssembly = pMD->GetLoaderModule()->GetDomainAssembly();
-    IfFailThrow(Attribute::ParseAttributeArgumentValues(
-        pData,
-        cData,
-        &caValueArrayFactory,
-        NULL,
-        0,
-        namedArgs,
-        lengthof(namedArgs),
-        domainAssembly));
-
-    // If the value isn't defined, then return without setting anything.
-    if (namedArgs[0].val.type.tag == SERIALIZATION_TYPE_UNDEFINED)
-        return false;
-
-    CorInfoCallConvExtension callConvLocal;
-    if (nativeCallableInternalData)
-    {
-        callConvLocal = (CorInfoCallConvExtension)(namedArgs[0].val.u4 << 8);
-    }
-    else
-    {
-        CallConvBuilder callConvBuilder;
-
-        CaValue* arrayOfTypes = &namedArgs[0].val;
-        for (ULONG i = 0; i < arrayOfTypes->arr.length; i++)
-        {
-            CaValue& typeNameValue = arrayOfTypes->arr[i];
-
-            if (!callConvBuilder.AddFullyQualifiedTypeName(typeNameValue.str.cbStr, typeNameValue.str.pStr))
-            {
-                // We found a second base calling convention.
-                return false;
-            }
-        }
-
-        callConvLocal = callConvBuilder.GetCurrentCallConv();
-        if (callConvLocal == CallConvBuilder::UnsetValue)
-        {
-            callConvLocal = MetaSig::GetDefaultUnmanagedCallingConvention();
-        }
-    }
-    *pCallConv = callConvLocal;
-    return true;
-}
 #endif // CROSSGEN_COMPILE
